requirements.txt
# Telegram Client
telethon==1.36.0

# AI Services
google-generativeai==0.8.3

# Translation
deep-translator==1.11.4

# PDF Processing
PyPDF2==3.0.1
pypdf==5.1.0

# Language Detection
langdetect==1.0.9

# Environment Management
python-dotenv==1.0.1

# HTTP Client
aiohttp==3.11.10

# Utilities
python-dateutil==2.9.0
